ingestion/requirements.txt
datetime
numpy
flask
pandas
mysql-connector-python